Update README.md
Aerial images have many distinct issues when compared
to natural images, such as a large number of dense in-
stances per image, and considerable object-scale fluctua-
tions. Another challenging issue with aerial images, is the
small ratio of foreground to background areas. For that
reason, and due to the class imbalance in aerial images
datasets, the over-presented classes will dominate the over-
all loss since they are generally easy to classify. To tackle
this issue, we propose to address this class imbalance by
swapping the binary cross-entropy loss in the RPN head
with a novel loss function. Our proposed loss, Salience
Biased Reduced Focal Loss (SBRFL), adds more weight
to hard examples; therefore, the detector will be protected
from the large number of easy negatives during training. In
addition, our proposed loss helps in increasing the contri-
butions of well-classified samples, which is not the case with
focal loss that focuses the attention on hard samples. Fur-
thermore, to evaluate the effectiveness of our loss function,
we designed a pipeline where we extract the features of the
images, generate saliency values and append them to into
our loss function. Our experimental results show that our
modified Faster-RCNN outperformed the baseline method
in the overall accuracy and the class-wise accuracy.